backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.36
PyMySQL==1.1.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pydantic==2.9.2
